requirements.txt
pytorch
tensorflow
tensorboardX
tensorflow
torchvision
onnx
caffe2
future